move trie.Entries to trie.go; go mod tidy; add runtime/types.go
