+++ runtime/types.go
@@ -1,0 +1,9 @@
+package runtime
+
+type Version struct {
+	Spec_name         []byte
+	Impl_name         []byte
+	Authoring_version int32
+	Spec_version      int32
+	Impl_version      int32
+}
