+++ runtime/wasmer.go
@@ -278,14 +278,6 @@ type Runtime struct {
 	trie *trie.Trie
 }
 
-type Version struct {
-	Spec_name         []byte
-	Impl_name         []byte
-	Authoring_version int32
-	Spec_version      int32
-	Impl_version      int32
-}
-
 func NewRuntime(fp string, t *trie.Trie) (*Runtime, error) {
 	// Reads the WebAssembly module as bytes.
 	bytes, err := wasm.ReadBytes(fp)
