--- trie/iterator_test.go
@@ -1,38 +1,0 @@
-package trie
-
-import (
-	"testing"
-)
-
-func TestEntries(t *testing.T) {
-	trie := newEmpty()
-
-	tests := []trieTest{
-		{key: []byte{0x01, 0x35}, value: []byte("pen")},
-		{key: []byte{0x01, 0x35, 0x79}, value: []byte("penguin")},
-		{key: []byte{0xf2}, value: []byte("feather")},
-		{key: []byte{0x09, 0xd3}, value: []byte("noot")},
-	}
-
-	for _, test := range tests {
-		err := trie.Put(test.key, test.value)
-		if err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	entries := trie.Entries()
-	for k, v := range entries {
-		t.Logf("key %x value %s", []byte(k), v)
-	}
-
-	if len(entries) != len(tests) {
-		t.Fatal("length of trie.Entries does not equal length of values put into trie")
-	}
-
-	for _, test := range tests {
-		if entries[string(test.key)] == nil {
-			t.Fatal("did not get entry in trie")
-		}
-	}
-}
