+++ trie/print.go
@@ -20,28 +20,6 @@ import (
 	"fmt"
 )
 
-// Entries returns all the key-value pairs in the trie as a map of keys to values
-func (t *Trie) Entries() map[string][]byte {
-	return t.entries(t.root, nil, make(map[string][]byte))
-}
-
-func (t *Trie) entries(current node, prefix []byte, kv map[string][]byte) map[string][]byte {
-	switch c := current.(type) {
-	case *branch:
-		if c.value != nil {
-			kv[string(nibblesToKeyLE(append(prefix, c.key...)))] = c.value
-		}
-		for i, child := range c.children {
-			t.entries(child, append(prefix, append(c.key, byte(i))...), kv)
-		}
-	case *leaf:
-		kv[string(nibblesToKeyLE(append(prefix, c.key...)))] = c.value
-		return kv
-	}
-
-	return kv
-}
-
 // Print prints the trie through pre-order traversal
 func (t *Trie) Print() {
 	fmt.Println("printing trie...")
