+++ trie/trie.go
@@ -71,6 +71,28 @@ func (t *Trie) Hash() (common.Hash, error) {
 	return common.Blake2bHash(encRoot)
 }
 
+// Entries returns all the key-value pairs in the trie as a map of keys to values
+func (t *Trie) Entries() map[string][]byte {
+	return t.entries(t.root, nil, make(map[string][]byte))
+}
+
+func (t *Trie) entries(current node, prefix []byte, kv map[string][]byte) map[string][]byte {
+	switch c := current.(type) {
+	case *branch:
+		if c.value != nil {
+			kv[string(nibblesToKeyLE(append(prefix, c.key...)))] = c.value
+		}
+		for i, child := range c.children {
+			t.entries(child, append(prefix, append(c.key, byte(i))...), kv)
+		}
+	case *leaf:
+		kv[string(nibblesToKeyLE(append(prefix, c.key...)))] = c.value
+		return kv
+	}
+
+	return kv
+}
+
 // Put inserts a key with value into the trie
 func (t *Trie) Put(key, value []byte) error {
 	if err := t.tryPut(key, value); err != nil {
